+++ silhouette-totp/build.sbt
@@ -1,0 +1,8 @@
+import Dependencies._
+
+libraryDependencies ++= Seq(
+  Library.googleAuth,
+  Library.Play.specs2 % Test
+)
+
+enablePlugins(Doc)
